# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GeminiConversation {
  id: number;
  title: string;
  mode: string;
  createdAt: string;
}

export interface MessageAttachment {
  objectPath: string;
  name: string;
  mimeType: string;
  size: number;
}

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  attachments?: MessageAttachment[];
  createdAt: string;
}

export interface CreateGeminiConversationBody {
  title: string;
  mode: string;
}

export interface UpdateGeminiConversationBody {
  title?: string;
  mode?: string;
}

export interface SendGeminiMessageBody {
  content: string;
  attachments?: MessageAttachment[];
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export interface GeminiConversationWithMessages {
  id: number;
  title: string;
  mode: string;
  createdAt: string;
  messages: GeminiMessage[];
}

export interface GeminiError {
  error: string;
}