# lib/api-zod/src/generated/types/geminiMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MessageAttachment } from "./messageAttachment";

export interface GeminiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  attachments?: MessageAttachment[];
  createdAt: Date;
}